[TAN-8637] Move useWidgetProjectId to the shared builder folder
