+++ front/app/components/ProjectPageBuilder/Widgets/ProjectBanner/index.tsx
@@ -14,6 +14,7 @@ import useProjectById from 'api/projects/useProjectById';
 import useLocalize from 'hooks/useLocalize';
 
 import { IMAGE_UPLOADING_EVENT } from 'components/admin/ContentBuilder/constants';
+import useWidgetProjectId from 'components/admin/ContentBuilder/useWidgetProjectId';
 import imageMessages from 'components/admin/ContentBuilder/Widgets/ImageMultiloc/messages';
 import ImageCropperContainer from 'components/admin/ImageCropper/Container';
 import {
@@ -31,7 +32,6 @@ import { usePermission } from 'utils/permissions';
 
 import LockedNote from '../LockedNote';
 import messages from '../messages';
-import useWidgetProjectId from '../useWidgetProjectId';
 
 import EmptyBanner from './EmptyBanner';
