+++ front/app/components/ProjectPageBuilder/regions.tsx
@@ -4,10 +4,11 @@ import { Box, colors } from '@citizenlab/cl2-component-library';
 import { Node, UserComponent, useEditor } from '@craftjs/core';
 import styled, { css } from 'styled-components';
 
+import useWidgetProjectId from 'components/admin/ContentBuilder/useWidgetProjectId';
+
 import { useParams } from 'utils/router';
 
 import CTABar from './CTABar';
-import useWidgetProjectId from './Widgets/useWidgetProjectId';
 
 type RegionProps = {
   children?: React.ReactNode;
